Implemented angle-between Had to add a manual tweak for the dicontinuity at the limit of ±PI, we force the result to always be positive in these cases. Besides making it simpler by only having one result for "180° away", this will really come into its own for the outline-drawing code as we can then just prefer to travel along lines with a negative bearing from us (i.e. "always turn left/ACW if available").

+++ triangberg_builder/include/triangberg_builder/types.hpp
@@ -20,7 +20,7 @@
 #include <cstddef>
 
 namespace com::saxbophone::triangberg {
-    typedef float Unit;
+    typedef double Unit;
 
     typedef Unit Degrees;
